Refactor avroToJsonParser
