+++ src/main/kotlin/insulator/di/LibModule.kt
@@ -3,8 +3,8 @@ package insulator.di
 import com.fasterxml.jackson.databind.ObjectMapper
 import insulator.lib.configuration.ConfigurationRepo
 import insulator.lib.configuration.model.Cluster
-import insulator.lib.jsonhelper.AvroToJsonConverter
 import insulator.lib.jsonhelper.JsonFormatter
+import insulator.lib.jsonhelper.avrotojson.AvroToJsonConverter
 import insulator.lib.jsonhelper.jsontoavro.FieldParser
 import insulator.lib.jsonhelper.jsontoavro.JsonToAvroConverter
 import insulator.lib.jsonhelper.jsontoavro.fieldparser.ComplexTypeParsersFactory
